# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/pages:
    get:
      tags:
        - Pages
      summary: List Pages
      description: Get a paginated list of Pages with optional filtering and sorting
      operationId: listPages
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Pages
      summary: Create Page
      description: Create a new Page
      operationId: createPage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePageRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/redirects:
    get:
      tags:
        - Redirects
      summary: List Redirects
      description: Get a paginated list of Redirects with optional filtering and sorting
      operationId: listRedirects
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedirectList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Redirects
      summary: Create Redirect
      description: Create a new Redirect
      operationId: createRedirect
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRedirectRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Redirect'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    Page:
      type: object
      required:
        - id
        - key
        - url
        - title
        - is_published
        - website_indexed
        - visibility
        - required_member_roles
        - metadata
      properties:
        id:
          type: string
          format: uuid
        key:
          type: string
          maxLength: 120
        website_id:
          nullable: true
          type: string
          format: uuid
        url:
          type: string
          maxLength: 500
        title:
          type: string
          maxLength: 200
        seo_name:
          nullable: true
          type: string
          maxLength: 120
        is_published:
          type: boolean
        date_publish:
          nullable: true
          type: string
          format: date-time
        website_indexed:
          type: boolean
        visibility:
          type: string
        required_member_roles:
          type: array
        forked_from:
          nullable: true
          type: string
          format: uuid
        forked_at:
          nullable: true
          type: string
          format: date-time
        forked_by:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    CreatePageRequest:
      type: object
      required:
        - key
        - url
        - title
        - is_published
        - website_indexed
        - visibility
        - required_member_roles
        - metadata
      properties:
        key:
          type: string
          maxLength: 120
        website_id:
          nullable: true
          type: string
          format: uuid
        url:
          type: string
          maxLength: 500
        title:
          type: string
          maxLength: 200
        seo_name:
          nullable: true
          type: string
          maxLength: 120
        is_published:
          type: boolean
        date_publish:
          nullable: true
          type: string
          format: date-time
        website_indexed:
          type: boolean
        visibility:
          type: string
        required_member_roles:
          type: array
        forked_from:
          nullable: true
          type: string
          format: uuid
        forked_at:
          nullable: true
          type: string
          format: date-time
        forked_by:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    PageList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Page'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    Redirect:
      type: object
      required:
        - id
        - website_id
        - url_from
        - redirect_type
        - metadata
      properties:
        id:
          type: string
          format: uuid
        website_id:
          type: string
          format: uuid
        url_from:
          type: string
          maxLength: 500
        redirect_type:
          type: string
        url_to:
          nullable: true
          type: string
          maxLength: 500
        description:
          nullable: true
          type: string
          maxLength: 500
        metadata:
          type: object
...
    CreateRedirectRequest:
      type: object
      required:
        - website_id
        - url_from
        - redirect_type
        - metadata
      properties:
        website_id:
          type: string
          format: uuid
        url_from:
          type: string
          maxLength: 500
        redirect_type:
          type: string
        url_to:
          nullable: true
          type: string
          maxLength: 500
        description:
          nullable: true
          type: string
          maxLength: 500
        metadata:
          type: object
...
    RedirectList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Redirect'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []